Migrate serialize tests
