+++ datafusion/proto/tests/cases/mod.rs
@@ -17,3 +17,4 @@
 
 mod roundtrip_logical_plan;
 mod roundtrip_physical_plan;
+mod serialize;
